+++ editor_patch/event.h
@@ -64,7 +64,8 @@ enum class AlpineDedEventID : int
     When_Captured,
     Set_Capture_Point_Owner,
     Owner_Gate,
-    Set_Gameplay_Rule
+    Set_Gameplay_Rule,
+    When_Round_Ends
 };
 
 constexpr int af_ded_event_to_int(AlpineDedEventID event_id) noexcept