+++ docs/CHANGELOG.md
@@ -8,6 +8,7 @@ Version 1.3.0 (TBD): Not yet released
 ### Minor features, changes, and enhancements
 [@GooberRF](https://github.com/GooberRF)
 - Add values to scope param for `Scope_Gate` event to test for team regardless of whether local player is spawned
+- Add `When_Round_Ends` event
 
 ### Bug fixes
 [@GooberRF](https://github.com/GooberRF)